Port ConstrData tests
