+++ plutarch-test/goldens/pisdata.constr.bench.golden
@@ -1,0 +1,8 @@
+sop.4wheeler {"exBudgetCPU":1494425,"exBudgetMemory":2760,"scriptSizeBytes":45}
+sop.2wheeler {"exBudgetCPU":956241,"exBudgetMemory":1896,"scriptSizeBytes":30}
+sop.immovable {"exBudgetCPU":298965,"exBudgetMemory":632,"scriptSizeBytes":12}
+prod.1 {"exBudgetCPU":1225333,"exBudgetMemory":2328,"scriptSizeBytes":40}
+prod.2 {"exBudgetCPU":1225333,"exBudgetMemory":2328,"scriptSizeBytes":68}
+enum.PA {"exBudgetCPU":298965,"exBudgetMemory":632,"scriptSizeBytes":12}
+enum.PB {"exBudgetCPU":298965,"exBudgetMemory":632,"scriptSizeBytes":12}
+pconstant-pcon-rel {"exBudgetCPU":29873,"exBudgetMemory":200,"scriptSizeBytes":22}
