+++ plutarch-test/goldens/pisdata.constr.uplc.eval.golden
@@ -1,0 +1,8 @@
+sop.4wheeler (program 1.0.0 #d8799f0205182a00ff)
+sop.2wheeler (program 1.0.0 #d87a9f0500ff)
+sop.immovable (program 1.0.0 #d87b80)
+prod.1 (program 1.0.0 #d8799f41ab4141410eff)
+prod.2 (program 1.0.0 #d8799fd8799f40ffd87a9fd8799fd8799f41abff00ffffd87b9fd8799fd8799f41daffffffff)
+enum.PA (program 1.0.0 #d87980)
+enum.PB (program 1.0.0 #d87a80)
+pconstant-pcon-rel (program 1.0.0 #d8799fd87a9f4101ffd87a80ff)
