+++ plutarch-test/goldens/pisdata.constr.uplc.golden
@@ -1,0 +1,8 @@
+sop.4wheeler (program 1.0.0 ((\i0 -> constrData 0 (i1 #02 (i1 #05 (i1 #182a (i1 #00 [  ]))))) (force mkCons)))
+sop.2wheeler (program 1.0.0 ((\i0 -> constrData 1 (i1 #05 (i1 #00 []))) (force mkCons)))
+sop.immovable (program 1.0.0 (constrData 2 []))
+prod.1 (program 1.0.0 ((\i0 -> constrData 0 (i1 #41ab (i1 #4141 (i1 #410e [  ])))) (force mkCons)))
+prod.2 (program 1.0.0 ((\i0 -> constrData 0 (i1 #d8799f40ff (i1 #d87a9fd8799fd8799f41abff00ffff (i1 #d87b9fd8799fd8799f41daffffff [  ])))) (force mkCons)))
+enum.PA (program 1.0.0 (constrData 0 []))
+enum.PB (program 1.0.0 (constrData 1 []))
+pconstant-pcon-rel (program 1.0.0 #d8799fd87a9f4101ffd87a80ff)
